[Misc] main2main 0519 (#9238)
### What this PR does / why we need it?

This PR updates vllm-ascend main2main validation to:

- vLLM version: `v0.20.2`
- vLLM main commit:
https://github.com/vllm-project/vllm/commit/1ac10f159a09897baada01b14b6a0dd6442aefd6
- vLLM diff:
https://github.com/vllm-project/vllm/compare/0d4d334eaa583b9c09aa4eb7538c22db99fd84b3...1ac10f159a09897baada01b14b6a0dd6442aefd6

Main upstream changes and vllm-ascend adaptations:

1. vLLM PR: https://github.com/vllm-project/vllm/pull/41775  
   `[Model Runner V2] FP32 gumbel sampling`

   Upstream changes:
   - Adds `use_fp64_gumbel` config / argument.
- Changes Gumbel sampling and rejection sampling paths to accept
`use_fp64`.
   - Makes FP32 Gumbel the default path and keeps FP64 optional.

   vllm-ascend adaptation:
- Update `vllm_ascend/worker/v2/sample/gumbel.py` to accept `use_fp64`.
- Update `vllm_ascend/worker/v2/spec_decode/rejection_sampler_utils.py`
to accept `use_fp64`.
- Raise `NotImplementedError` for `use_fp64=True` on NPU because the
current NPU Triton path does not support FP64 Gumbel / rejection
sampling.

2. vLLM PR: https://github.com/vllm-project/vllm/pull/41162  
   `[Model Runner V2] Rebuild attn metadata between draft decode steps`

   Upstream changes:
- Adds `output_processed_logits` / `output_processed_logits_col`
protocol for EAGLE draft sampling.
- Uses `output_processed_logits_col` to select the current speculative
draft step when writing draft logits.

   vllm-ascend adaptation:
- Add `output_processed_logits` and `output_processed_logits_col`
support in NPU `gumbel_sample`.
- Store processed logits before adding Gumbel noise so rejection
sampling can consume the draft logits.


3. vLLM PR: https://github.com/vllm-project/vllm/pull/42692  
   `[Bugfix] DFlash FP8 KV-Cache`

   Upstream changes:
- Changes `SpecDecodeBaseProposer.token_arange_np` from default NumPy
integer dtype to `np.int32`.

   vllm-ascend adaptation:
- Update `vllm_ascend/spec_decode/eagle_proposer.py` to use
`np.arange(..., dtype=np.int32)`.
- Keep the Ascend-specific `max_num_tokens + 1` behavior for
`query_start_loc_cpu[:batch_size + 1]`.

4. No direct upstream vLLM PR

   vllm-ascend adaptation:
   - Fix `OP_LOGE` / `OP_LOGW` format warnings in:
     - `csrc/moe/chunk_fwd_o/tiling_base/error_log.h`
     - `csrc/moe/chunk_gated_delta_rule_fwd_h/tiling_base/error_log.h`

   Reason:
- The old macros passed `std::string` to `%s`, which fails when the
current build treats format warnings as errors.


### Does this PR introduce _any_ user-facing change?

No.

This PR is for main2main compatibility validation and internal NPU
sampling / speculative decoding adaptation.

One behavior to note: if users explicitly enable `--use-fp64-gumbel` on
NPU, vllm-ascend will raise `NotImplementedError`.

### How was this patch tested?

- vLLM version: `v0.20.2`
- vLLM main commit:
https://github.com/vllm-project/vllm/commit/1ac10f159a09897baada01b14b6a0dd6442aefd6
- vLLM diff:
https://github.com/vllm-project/vllm/compare/0d4d334eaa583b9c09aa4eb7538c22db99fd84b3...1ac10f159a09897baada01b14b6a0dd6442aefd6
- Validation focus:
  - NPU Gumbel sampling
  - EAGLE / MTP speculative decoding
  - rejection sampling
  - custom op build

---------

+++ vllm_ascend/spec_decode/llm_base_proposer.py
@@ -182,7 +182,7 @@ class AscendSpecDecodeBaseProposer(SpecDecodeBaseProposer):
             # RoPE need (max_num_tokens,)
             self.positions = torch.zeros(self.max_num_tokens, dtype=torch.int32, device=device)
 
-        self.token_arange_np = np.arange(self.max_num_tokens + 1)
+        self.token_arange_np = np.arange(self.max_num_tokens + 1, dtype=np.int32)
         self.enable_enpu = self.runner.enable_enpu
         self.use_eagle = self.runner.use_eagle
